Append Field to the form field class names
