+++ administrator/components/com_content/Field/Modal/ArticleField.php
@@ -1,31 +1,32 @@
 <?php
 /**
  * @package     Joomla.Administrator
- * @subpackage  com_categories
+ * @subpackage  com_content
  *
  * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
-namespace Joomla\Component\Categories\Administrator\Field\Modal;
+
+namespace Joomla\Component\Content\Administrator\Field\Modal;
 
 defined('JPATH_BASE') or die;
 
 use Joomla\CMS\Form\FormField;
 
 /**
- * Supports a modal category picker.
+ * Supports a modal article picker.
  *
- * @since  3.1
+ * @since  1.6
  */
-class Category extends FormField
+class ArticleField extends FormField
 {
 	/**
 	 * The form field type.
 	 *
-	 * @var     string
-	 * @since   1.6
+	 * @var    string
+	 * @since  1.6
 	 */
-	protected $type = 'Modal_Category';
+	protected $type = 'Modal_Article';
 
 	/**
 	 * Method to get the field input markup.
@@ -36,28 +37,19 @@ class Category extends FormField
 	 */
 	protected function getInput()
 	{
-		if ($this->element['extension'])
-		{
-			$extension = (string) $this->element['extension'];
-		}
-		else
-		{
-			$extension = (string) \JFactory::getApplication()->input->get('extension', 'com_content');
-		}
-
 		$allowNew    = ((string) $this->element['new'] == 'true');
 		$allowEdit   = ((string) $this->element['edit'] == 'true');
 		$allowClear  = ((string) $this->element['clear'] != 'false');
 		$allowSelect = ((string) $this->element['select'] != 'false');
 
-		// Load language.
-		\JFactory::getLanguage()->load('com_categories', JPATH_ADMINISTRATOR);
+		// Load language
+		\JFactory::getLanguage()->load('com_content', JPATH_ADMINISTRATOR);
 
-		// The active category id field.
+		// The active article id field.
 		$value = (int) $this->value > 0 ? (int) $this->value : '';
 
 		// Create the modal id.
-		$modalId = 'Category_' . $this->id;
+		$modalId = 'Article_' . $this->id;
 
 		// Add the modal field script to the document head.
 		\JHtml::_('jquery.framework');
@@ -76,8 +68,8 @@ class Category extends FormField
 			if (!isset($scriptSelect[$this->id]))
 			{
 				\JFactory::getDocument()->addScriptDeclaration("
-				function jSelectCategory_" . $this->id . "(id, title, object) {
-					window.processModalSelect('Category', '" . $this->id . "', id, title, '', object);
+				function jSelectArticle_" . $this->id . "(id, title, catid, object, url, language) {
+					window.processModalSelect('Article', '" . $this->id . "', id, title, catid, object, url, language);
 				}
 				");
 
@@ -86,29 +78,30 @@ class Category extends FormField
 		}
 
 		// Setup variables for display.
-		$linkCategories = 'index.php?option=com_categories&amp;view=categories&amp;layout=modal&amp;tmpl=component&amp;' . \JSession::getFormToken() . '=1'
-			. '&amp;extension=' . $extension;
-		$linkCategory  = 'index.php?option=com_categories&amp;view=category&amp;layout=modal&amp;tmpl=component&amp;' . \JSession::getFormToken() . '=1'
-			. '&amp;extension=' . $extension;
-		$modalTitle    = \JText::_('COM_CATEGORIES_CHANGE_CATEGORY');
+		$linkArticles = 'index.php?option=com_content&amp;view=articles&amp;layout=modal&amp;tmpl=component&amp;' . \JSession::getFormToken() . '=1';
+		$linkArticle  = 'index.php?option=com_content&amp;view=article&amp;layout=modal&amp;tmpl=component&amp;' . \JSession::getFormToken() . '=1';
 
 		if (isset($this->element['language']))
 		{
-			$linkCategories .= '&amp;forcedLanguage=' . $this->element['language'];
-			$linkCategory   .= '&amp;forcedLanguage=' . $this->element['language'];
-			$modalTitle     .= ' &#8212; ' . $this->element['label'];
+			$linkArticles .= '&amp;forcedLanguage=' . $this->element['language'];
+			$linkArticle  .= '&amp;forcedLanguage=' . $this->element['language'];
+			$modalTitle    = \JText::_('COM_CONTENT_CHANGE_ARTICLE') . ' &#8212; ' . $this->element['label'];
+		}
+		else
+		{
+			$modalTitle    = \JText::_('COM_CONTENT_CHANGE_ARTICLE');
 		}
 
-		$urlSelect = $linkCategories . '&amp;function=jSelectCategory_' . $this->id;
-		$urlEdit   = $linkCategory . '&amp;task=category.edit&amp;id=\' + document.getElementById("' . $this->id . '_id").value + \'';
-		$urlNew    = $linkCategory . '&amp;task=category.add';
+		$urlSelect = $linkArticles . '&amp;function=jSelectArticle_' . $this->id;
+		$urlEdit   = $linkArticle . '&amp;task=article.edit&amp;id=\' + document.getElementById("' . $this->id . '_id").value + \'';
+		$urlNew    = $linkArticle . '&amp;task=article.add';
 
 		if ($value)
 		{
 			$db    = \JFactory::getDbo();
 			$query = $db->getQuery(true)
 				->select($db->quoteName('title'))
-				->from($db->quoteName('#__categories'))
+				->from($db->quoteName('#__content'))
 				->where($db->quoteName('id') . ' = ' . (int) $value);
 			$db->setQuery($query);
 
@@ -122,9 +115,9 @@ class Category extends FormField
 			}
 		}
 
-		$title = empty($title) ? \JText::_('COM_CATEGORIES_SELECT_A_CATEGORY') : htmlspecialchars($title, ENT_QUOTES, 'UTF-8');
+		$title = empty($title) ? \JText::_('COM_CONTENT_SELECT_AN_ARTICLE') : htmlspecialchars($title, ENT_QUOTES, 'UTF-8');
 
-		// The current category display field.
+		// The current article display field.
 		$html  = '';
 		if ($allowSelect || $allowNew || $allowEdit || $allowClear)
 		{
@@ -138,7 +131,7 @@ class Category extends FormField
 			$html .= '<span class="input-group-btn">';
 		}
 
-		// Select category button.
+		// Select article button
 		if ($allowSelect)
 		{
 			$html .= '<a'
@@ -147,12 +140,12 @@ class Category extends FormField
 				. ' data-toggle="modal"'
 				. ' role="button"'
 				. ' href="#ModalSelect' . $modalId . '"'
-				. ' title="' . \JHtml::tooltipText('COM_CATEGORIES_CHANGE_CATEGORY') . '">'
+				. ' title="' . \JHtml::tooltipText('COM_CONTENT_CHANGE_ARTICLE') . '">'
 				. '<span class="icon-file" aria-hidden="true"></span> ' . \JText::_('JSELECT')
 				. '</a>';
 		}
 
-		// New category button.
+		// New article button
 		if ($allowNew)
 		{
 			$html .= '<a'
@@ -161,12 +154,12 @@ class Category extends FormField
 				. ' data-toggle="modal"'
 				. ' role="button"'
 				. ' href="#ModalNew' . $modalId . '"'
-				. ' title="' . \JHtml::tooltipText('COM_CATEGORIES_NEW_CATEGORY') . '">'
+				. ' title="' . \JHtml::tooltipText('COM_CONTENT_NEW_ARTICLE') . '">'
 				. '<span class="icon-new" aria-hidden="true"></span> ' . \JText::_('JACTION_CREATE')
 				. '</a>';
 		}
 
-		// Edit category button.
+		// Edit article button
 		if ($allowEdit)
 		{
 			$html .= '<a'
@@ -175,12 +168,12 @@ class Category extends FormField
 				. ' data-toggle="modal"'
 				. ' role="button"'
 				. ' href="#ModalEdit' . $modalId . '"'
-				. ' title="' . \JHtml::tooltipText('COM_CATEGORIES_EDIT_CATEGORY') . '">'
+				. ' title="' . \JHtml::tooltipText('COM_CONTENT_EDIT_ARTICLE') . '">'
 				. '<span class="icon-edit" aria-hidden="true"></span> ' . \JText::_('JACTION_EDIT')
 				. '</a>';
 		}
 
-		// Clear category button.
+		// Clear article button
 		if ($allowClear)
 		{
 			$html .= '<a'
@@ -197,7 +190,7 @@ class Category extends FormField
 			$html .= '</span></span>';
 		}
 
-		// Select category modal.
+		// Select article modal
 		if ($allowSelect)
 		{
 			$html .= \JHtml::_(
@@ -216,14 +209,14 @@ class Category extends FormField
 			);
 		}
 
-		// New category modal.
+		// New article modal
 		if ($allowNew)
 		{
 			$html .= \JHtml::_(
 				'bootstrap.renderModal',
 				'ModalNew' . $modalId,
 				array(
-					'title'       => \JText::_('COM_CATEGORIES_NEW_CATEGORY'),
+					'title'       => \JText::_('COM_CONTENT_NEW_ARTICLE'),
 					'backdrop'    => 'static',
 					'keyboard'    => false,
 					'closeButton' => false,
@@ -233,26 +226,26 @@ class Category extends FormField
 					'bodyHeight'  => 70,
 					'modalWidth'  => 80,
 					'footer'      => '<a role="button" class="btn btn-secondary" aria-hidden="true"'
-							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'category\', \'cancel\', \'item-form\'); return false;">'
+							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'article\', \'cancel\', \'item-form\'); return false;">'
 							. \JText::_('JLIB_HTML_BEHAVIOR_CLOSE') . '</a>'
 							. '<a role="button" class="btn btn-primary" aria-hidden="true"'
-							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'category\', \'save\', \'item-form\'); return false;">'
+							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'article\', \'save\', \'item-form\'); return false;">'
 							. \JText::_('JSAVE') . '</a>'
 							. '<a role="button" class="btn btn-success" aria-hidden="true"'
-							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'category\', \'apply\', \'item-form\'); return false;">'
+							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'add\', \'article\', \'apply\', \'item-form\'); return false;">'
 							. \JText::_('JAPPLY') . '</a>',
 				)
 			);
 		}
 
-		// Edit category modal.
+		// Edit article modal
 		if ($allowEdit)
 		{
 			$html .= \JHtml::_(
 				'bootstrap.renderModal',
 				'ModalEdit' . $modalId,
 				array(
-					'title'       => \JText::_('COM_CATEGORIES_EDIT_CATEGORY'),
+					'title'       => \JText::_('COM_CONTENT_EDIT_ARTICLE'),
 					'backdrop'    => 'static',
 					'keyboard'    => false,
 					'closeButton' => false,
@@ -262,23 +255,23 @@ class Category extends FormField
 					'bodyHeight'  => 70,
 					'modalWidth'  => 80,
 					'footer'      => '<a role="button" class="btn btn-secondary" aria-hidden="true"'
-							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'category\', \'cancel\', \'item-form\'); return false;">'
+							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'article\', \'cancel\', \'item-form\'); return false;">'
 							. \JText::_('JLIB_HTML_BEHAVIOR_CLOSE') . '</a>'
 							. '<a role="button" class="btn btn-primary" aria-hidden="true"'
-							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'category\', \'save\', \'item-form\'); return false;">'
+							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'article\', \'save\', \'item-form\'); return false;">'
 							. \JText::_('JSAVE') . '</a>'
 							. '<a role="button" class="btn btn-success" aria-hidden="true"'
-							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'category\', \'apply\', \'item-form\'); return false;">'
+							. ' onclick="window.processModalEdit(this, \'' . $this->id . '\', \'edit\', \'article\', \'apply\', \'item-form\'); return false;">'
 							. \JText::_('JAPPLY') . '</a>',
 				)
 			);
 		}
 
-		// Note: class='required' for client side validation
+		// Note: class='required' for client side validation.
 		$class = $this->required ? ' class="required modal-value"' : '';
 
-		$html .= '<input type="hidden" id="' . $this->id . '_id"' . $class . ' data-required="' . (int) $this->required . '" name="' . $this->name
-			. '" data-text="' . htmlspecialchars(\JText::_('COM_CATEGORIES_SELECT_A_CATEGORY', true), ENT_COMPAT, 'UTF-8') . '" value="' . $value . '">';
+		$html .= '<input type="hidden" id="' . $this->id . '_id" ' . $class . ' data-required="' . (int) $this->required . '" name="' . $this->name
+			. '" data-text="' . htmlspecialchars(\JText::_('COM_CONTENT_SELECT_AN_ARTICLE', true), ENT_COMPAT, 'UTF-8') . '" value="' . $value . '">';
 
 		return $html;
 	}
@@ -288,7 +281,7 @@ class Category extends FormField
 	 *
 	 * @return  string  The field label markup.
 	 *
-	 * @since   3.7.0
+	 * @since   3.4
 	 */
 	protected function getLabel()
 	{
